+++ scrapy/commands/parse.py
@@ -1,16 +1,32 @@
+import argparse
 import functools
 import inspect
 import json
 import logging
-from typing import Dict
+from types import CoroutineType
+from typing import (
+    Any,
+    AsyncGenerator,
+    Callable,
+    Dict,
+    Iterable,
+    List,
+    Optional,
+    Tuple,
+    TypeVar,
+    Union,
+    overload,
+)
 
 from itemadapter import ItemAdapter, is_item
-from twisted.internet.defer import maybeDeferred
+from twisted.internet.defer import Deferred, maybeDeferred
+from twisted.python.failure import Failure
 from w3lib.url import is_url
 
 from scrapy.commands import BaseRunSpiderCommand
 from scrapy.exceptions import UsageError
-from scrapy.http import Request
+from scrapy.http import Request, Response
+from scrapy.spiders import Spider
 from scrapy.utils import display
 from scrapy.utils.asyncgen import collect_asyncgen
 from scrapy.utils.defer import aiter_errback, deferred_from_coro
@@ -20,24 +36,26 @@ from scrapy.utils.spider import spidercls_for_request
 
 logger = logging.getLogger(__name__)
 
+_T = TypeVar("_T")
+
 
 class Command(BaseRunSpiderCommand):
     requires_project = True
 
     spider = None
-    items: Dict[int, list] = {}
-    requests: Dict[int, list] = {}
+    items: Dict[int, List[Any]] = {}
+    requests: Dict[int, List[Request]] = {}
 
     first_response = None
 
-    def syntax(self):
+    def syntax(self) -> str:
         return "[options] <url>"
 
-    def short_desc(self):
+    def short_desc(self) -> str:
         return "Parse URL (using its spider) and print the results"
 
-    def add_options(self, parser):
-        BaseRunSpiderCommand.add_options(self, parser)
+    def add_options(self, parser: argparse.ArgumentParser) -> None:
+        super().add_options(parser)
         parser.add_argument(
             "--spider",
             dest="spider",
@@ -106,7 +124,7 @@ class Command(BaseRunSpiderCommand):
         )
 
     @property
-    def max_level(self):
+    def max_level(self) -> int:
         max_items, max_requests = 0, 0
         if self.items:
             max_items = max(self.items)
@@ -114,13 +132,21 @@ class Command(BaseRunSpiderCommand):
             max_requests = max(self.requests)
         return max(max_items, max_requests)
 
-    def handle_exception(self, _failure):
+    def handle_exception(self, _failure: Failure) -> None:
         logger.error(
             "An error is caught while iterating the async iterable",
             exc_info=failure_to_exc_info(_failure),
         )
 
-    def iterate_spider_output(self, result):
+    @overload
+    def iterate_spider_output(
+        self, result: Union[AsyncGenerator, CoroutineType]
+    ) -> Deferred: ...
+
+    @overload
+    def iterate_spider_output(self, result: _T) -> Iterable: ...
+
+    def iterate_spider_output(self, result: Any) -> Union[Iterable, Deferred]:
         if inspect.isasyncgen(result):
             d = deferred_from_coro(
                 collect_asyncgen(aiter_errback(result, self.handle_exception))
@@ -133,15 +159,15 @@ class Command(BaseRunSpiderCommand):
             return d
         return arg_to_iter(deferred_from_coro(result))
 
-    def add_items(self, lvl, new_items):
+    def add_items(self, lvl: int, new_items: List[Any]) -> None:
         old_items = self.items.get(lvl, [])
         self.items[lvl] = old_items + new_items
 
-    def add_requests(self, lvl, new_reqs):
+    def add_requests(self, lvl: int, new_reqs: List[Request]) -> None:
         old_reqs = self.requests.get(lvl, [])
         self.requests[lvl] = old_reqs + new_reqs
 
-    def print_items(self, lvl=None, colour=True):
+    def print_items(self, lvl: Optional[int] = None, colour: bool = True) -> None:
         if lvl is None:
             items = [item for lst in self.items.values() for item in lst]
         else:
@@ -150,7 +176,7 @@ class Command(BaseRunSpiderCommand):
         print("# Scraped Items ", "-" * 60)
         display.pprint([ItemAdapter(x).asdict() for x in items], colorize=colour)
 
-    def print_requests(self, lvl=None, colour=True):
+    def print_requests(self, lvl: Optional[int] = None, colour: bool = True) -> None:
         if lvl is None:
             if self.requests:
                 requests = self.requests[max(self.requests)]
@@ -162,7 +188,7 @@ class Command(BaseRunSpiderCommand):
         print("# Requests ", "-" * 65)
         display.pprint(requests, colorize=colour)
 
-    def print_results(self, opts):
+    def print_results(self, opts: argparse.Namespace) -> None:
         colour = not opts.nocolour
 
         if opts.verbose:
@@ -179,7 +205,14 @@ class Command(BaseRunSpiderCommand):
             if not opts.nolinks:
                 self.print_requests(colour=colour)
 
-    def _get_items_and_requests(self, spider_output, opts, depth, spider, callback):
+    def _get_items_and_requests(
+        self,
+        spider_output: Iterable[Any],
+        opts: argparse.Namespace,
+        depth: int,
+        spider: Spider,
+        callback: Callable,
+    ) -> Tuple[List[Any], List[Request], argparse.Namespace, int, Spider, Callable]:
         items, requests = [], []
         for x in spider_output:
             if is_item(x):
@@ -188,14 +221,21 @@ class Command(BaseRunSpiderCommand):
                 requests.append(x)
         return items, requests, opts, depth, spider, callback
 
-    def run_callback(self, response, callback, cb_kwargs=None):
+    def run_callback(
+        self,
+        response: Response,
+        callback: Callable,
+        cb_kwargs: Optional[Dict[str, Any]] = None,
+    ) -> Deferred:
         cb_kwargs = cb_kwargs or {}
         d = maybeDeferred(self.iterate_spider_output, callback(response, **cb_kwargs))
         return d
 
-    def get_callback_from_rules(self, spider, response):
+    def get_callback_from_rules(
+        self, spider: Spider, response: Response
+    ) -> Union[Callable, str, None]:
         if getattr(spider, "rules", None):
-            for rule in spider.rules:
+            for rule in spider.rules:  # type: ignore[attr-defined]
                 if rule.link_extractor.matches(response.url):
                     return rule.callback or "parse"
         else:
@@ -204,8 +244,10 @@ class Command(BaseRunSpiderCommand):
                 "please specify a callback to use for parsing",
                 {"spider": spider.name},
             )
+        return None
 
-    def set_spidercls(self, url, opts):
+    def set_spidercls(self, url: str, opts: argparse.Namespace) -> None:
+        assert self.crawler_process
         spider_loader = self.crawler_process.spider_loader
         if opts.spider:
             try:
@@ -219,13 +261,14 @@ class Command(BaseRunSpiderCommand):
             if not self.spidercls:
                 logger.error("Unable to find spider for: %(url)s", {"url": url})
 
-        def _start_requests(spider):
+        def _start_requests(spider: Spider) -> Iterable[Request]:
             yield self.prepare_request(spider, Request(url), opts)
 
         if self.spidercls:
             self.spidercls.start_requests = _start_requests
 
-    def start_parsing(self, url, opts):
+    def start_parsing(self, url: str, opts: argparse.Namespace) -> None:
+        assert self.crawler_process
         self.crawler_process.crawl(self.spidercls, **opts.spargs)
         self.pcrawler = list(self.crawler_process.crawlers)[0]
         self.crawler_process.start()
@@ -233,7 +276,12 @@ class Command(BaseRunSpiderCommand):
         if not self.first_response:
             logger.error("No response downloaded for: %(url)s", {"url": url})
 
-    def scraped_data(self, args):
+    def scraped_data(
+        self,
+        args: Tuple[
+            List[Any], List[Request], argparse.Namespace, int, Spider, Callable
+        ],
+    ) -> List[Any]:
         items, requests, opts, depth, spider, callback = args
         if opts.pipelines:
             itemproc = self.pcrawler.engine.scraper.itemproc
@@ -252,8 +300,14 @@ class Command(BaseRunSpiderCommand):
 
         return scraped_data
 
-    def _get_callback(self, *, spider, opts, response=None):
-        cb = None
+    def _get_callback(
+        self,
+        *,
+        spider: Spider,
+        opts: argparse.Namespace,
+        response: Optional[Response] = None,
+    ) -> Callable:
+        cb: Union[str, Callable, None] = None
         if response:
             cb = response.meta["_callback"]
         if not cb:
@@ -270,6 +324,7 @@ class Command(BaseRunSpiderCommand):
                 cb = "parse"
 
         if not callable(cb):
+            assert cb is not None
             cb_method = getattr(spider, cb, None)
             if callable(cb_method):
                 cb = cb_method
@@ -277,10 +332,13 @@ class Command(BaseRunSpiderCommand):
                 raise ValueError(
                     f"Cannot find callback {cb!r} in spider: {spider.name}"
                 )
+        assert callable(cb)
         return cb
 
-    def prepare_request(self, spider, request, opts):
-        def callback(response, **cb_kwargs):
+    def prepare_request(
+        self, spider: Spider, request: Request, opts: argparse.Namespace
+    ) -> Request:
+        def callback(response: Response, **cb_kwargs: Any) -> Deferred:
             # memorize first request
             if not self.first_response:
                 self.first_response = response
@@ -288,7 +346,7 @@ class Command(BaseRunSpiderCommand):
             cb = self._get_callback(spider=spider, opts=opts, response=response)
 
             # parse items and requests
-            depth = response.meta["_depth"]
+            depth: int = response.meta["_depth"]
 
             d = self.run_callback(response, cb, cb_kwargs)
             d.addCallback(self._get_items_and_requests, opts, depth, spider, callback)
@@ -311,13 +369,13 @@ class Command(BaseRunSpiderCommand):
         request.callback = callback
         return request
 
-    def process_options(self, args, opts):
-        BaseRunSpiderCommand.process_options(self, args, opts)
+    def process_options(self, args: List[str], opts: argparse.Namespace) -> None:
+        super().process_options(args, opts)
 
         self.process_request_meta(opts)
         self.process_request_cb_kwargs(opts)
 
-    def process_request_meta(self, opts):
+    def process_request_meta(self, opts: argparse.Namespace) -> None:
         if opts.meta:
             try:
                 opts.meta = json.loads(opts.meta)
@@ -328,7 +386,7 @@ class Command(BaseRunSpiderCommand):
                     print_help=False,
                 )
 
-    def process_request_cb_kwargs(self, opts):
+    def process_request_cb_kwargs(self, opts: argparse.Namespace) -> None:
         if opts.cbkwargs:
             try:
                 opts.cbkwargs = json.loads(opts.cbkwargs)
@@ -339,7 +397,7 @@ class Command(BaseRunSpiderCommand):
                     print_help=False,
                 )
 
-    def run(self, args, opts):
+    def run(self, args: List[str], opts: argparse.Namespace) -> None:
         # parse arguments
         if not len(args) == 1 or not is_url(args[0]):
             raise UsageError()